+++ binance/FFT.py
@@ -187,7 +187,7 @@ class FFT(IStrategy):
     ###################################
 
 
-    def model(self, a: np.ndarray) -> np.float:
+    def model(self, a: np.ndarray) -> float:
         #must return scalar, so just calculate prediction and take last value
 
         # scale the data
@@ -229,7 +229,7 @@ class FFT(IStrategy):
 
         return model
 
-    def scaledModel(self, a: np.ndarray) -> np.float:
+    def scaledModel(self, a: np.ndarray) -> float:
 
         # scale the data
         standardized = a.copy()
@@ -244,7 +244,7 @@ class FFT(IStrategy):
         length = len(model)
         return model[length-1]
 
-    def scaledData(self, a: np.ndarray) -> np.float:
+    def scaledData(self, a: np.ndarray) -> float:
 
         # scale the data
         standardized = a.copy()
@@ -256,7 +256,7 @@ class FFT(IStrategy):
         length = len(scaled)
         return scaled.ravel()[length-1]
 
-    def predict(self, a: np.ndarray) -> np.float:
+    def predict(self, a: np.ndarray) -> float:
         #must return scalar, so just calculate prediction and take last value
         npredict = self.fft_lookahead
         # y = self.fourierExtrapolation(np.array(a), 0)
