+++ binanceus/LSTM.py
@@ -675,7 +675,7 @@ class LSTM(IStrategy):
         print("runs:{} predictions:{}".format(nruns, len(predictions)))
         return predictions
 
-    def roll_get_prediction(self, col) -> np.float:
+    def roll_get_prediction(self, col) -> float:
         # must return scalar, so just calculate prediction and take last value
 
         # print(col.index)
@@ -693,7 +693,7 @@ class LSTM(IStrategy):
 
 
     # returns (rolling) smoothed version of input column
-    def roll_smooth(self, col) -> np.float:
+    def roll_smooth(self, col) -> float:
         # must return scalar, so just calculate prediction and take last value
 
         smooth = gaussian_filter1d(col, 4)
@@ -705,7 +705,7 @@ class LSTM(IStrategy):
         else:
             return col[len(col) - 1]
 
-    def roll_strong_smooth(self, col) -> np.float:
+    def roll_strong_smooth(self, col) -> float:
         # must return scalar, so just calculate prediction and take last value
 
         smooth = gaussian_filter1d(col, 24)
