+++ kucoin/DWT_Leveraged2.py
@@ -272,7 +272,7 @@ class DWT_Leveraged2(IStrategy):
 
         return model
 
-    def model(self, a: np.ndarray) -> np.float:
+    def model(self, a: np.ndarray) -> float:
         #must return scalar, so just calculate prediction and take last value
         # model = self.dwtModel(np.array(a))
